perf: move assign subscription modal to assignSub folder
